ECOM-main/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.12.0
django-parler==2.2
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
python-gettext==4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
whitenoise==5.3.0
